FIX: Remove line overhang
Previously, the line stretched all the way to the leftmost part of the child
box. This is ugly because the conclusion takes up only a bit of the middle, so
the line would extend past the child conclusion. The calculations now take
care of that so that the line only draws to the edges of the formulae.