IMP: Remove huge unnecessary gutters between branches
Had a display problem where, because we were ensuring constant spacing
of the children, huge gaps ocurred when they were not all of the same
size. This now ensures that children boxes are close to each other
(note this is not the same thing as minimizing the space between them
as, because the children are roughly triangular, you can have gaps
when one is taller than eachother, illustration:

    ___________
	\         /
	 \       /
	  \     /
	   \   /    ___
	    \ /     \ /
		 -       -

Would be produced rather than

    ___________
	\         /
	 \       /
	  \     /
	   \   / ___
	    \ /  \ /
		 -    -

but it is not clear we should prefer the latter).